celeritas/requirements.txt
aiohttp==3.9.5
base58==2.1.1
construct==2.10.68
httpx==0.27.0
jupiter_python_sdk==0.0.2.0
pymongo==4.8.0
python-telegram-bot==21.3
Requests==2.32.3
solana==0.34.2
solders==0.21.0
spl==0.0.1.dev0
websockets==10.4
